homework 3.4 complete v.2.2

+++ src/main/java/ru/hogwarts/school/service/impl/FacultyServiceImpl.java
@@ -1,6 +1,7 @@
 package ru.hogwarts.school.service.impl;
 
 import org.springframework.stereotype.Service;
+import ru.hogwarts.school.exception.FacultyNotFoundException;
 import ru.hogwarts.school.model.Faculty;
 import ru.hogwarts.school.model.Student;
 import ru.hogwarts.school.repositories.FacultyRepository;
@@ -13,10 +14,11 @@ import java.util.Collection;
 public class FacultyServiceImpl implements FacultyService {
 
     private final FacultyRepository facultyRepository;
-    private StudentRepository studentRepository;
+    private final StudentRepository studentRepository;
 
-    public FacultyServiceImpl(FacultyRepository facultyRepository) {
+    public FacultyServiceImpl(FacultyRepository facultyRepository, StudentRepository studentRepository) {
         this.facultyRepository = facultyRepository;
+        this.studentRepository = studentRepository;
     }
 
     @Override
@@ -25,17 +27,24 @@ public class FacultyServiceImpl implements FacultyService {
     }
 
     @Override
-    public Faculty findFaculty(long id) {
-        return facultyRepository.getById(id);
+    public Faculty findFaculty(Long id) {
+        return facultyRepository.findById(id).orElseThrow(() ->
+                new FacultyNotFoundException(id));
     }
 
     @Override
     public void deleteFaculty(long id) {
+        if (!facultyRepository.existsById(id)) {
+            throw new FacultyNotFoundException(id);
+        }
         facultyRepository.deleteById(id);
     }
 
     @Override
     public Faculty editFaculty(Faculty faculty) {
+        if (!facultyRepository.existsById(faculty.getId())) {
+            throw new FacultyNotFoundException(faculty.getId());
+        }
         return facultyRepository.save(faculty);
     }
 
@@ -51,8 +60,8 @@ public class FacultyServiceImpl implements FacultyService {
 
     @Override
     public Collection<Student> getStudentsByFaculty(long id) {
-        Faculty faculty = facultyRepository.findAllById(id);
-        return studentRepository.findAll(faculty);
+        Faculty faculty = facultyRepository.findById(id)
+                .orElseThrow(() -> new FacultyNotFoundException(id));
+        return studentRepository.findAllByFaculty(faculty);
     }
 }
-
